refactor: extract ProfileInput component
